Enum noodles_vcf::header::info::key::Key

pub enum Key {

    AncestralAllele,
    AlleleCount,
    TotalReadDepths,
    ForwardStrandReadDepths,
    ReverseStrandReadDepths,
    AlleleFrequencies,
    TotalAlleleCount,
    BaseQuality,
    Cigar,
    IsInDbSnp,
    TotalDepth,
    IsInHapMap2,
    IsInHapMap3,
    MappingQuality,
    ZeroMappingQualityCount,
    SamplesWithDataCount,
    StrandBias,
    IsSomaticMutation,
    IsValidated,
    IsIn1000Genomes,
    IsImprecise,
    IsNovel,
    EndPosition,
    SvType,
    SvLengths,
    PositionConfidenceIntervals,
    EndConfidenceIntervals,
    MicrohomologyLengths,
    MicrohomologySequences,
    BreakpointIds,
    MobileElementInfo,
    MobileElementTransductionInfo,
    DbvId,
    DbVarId,
    DbRipId,
    MateBreakendIds,
    PartnerBreakendId,
    BreakendEventId,
    BreakendConfidenceIntervals,
    AdjacentReadDepths,
    BreakendCopyNumber,
    AdjacentCopyNumber,
    CopyNumberConfidenceIntervals,
    AdjacentCopyNumberConfidenceIntervals,
    Other(String),

}

A VCF header info key.

Variants

AncestralAllele

Ancestral allele (AA).

AlleleCount

Allele count in genotypes, for each ALT allele, in the same order as listed (AC).

TotalReadDepths

Total read depth for each allele (AD).

ForwardStrandReadDepths

Read depth for each allele on the forward strand (ADF).

ReverseStrandReadDepths

Read depth for each allele on the reverse strand (ADR).

AlleleFrequencies

Allele frequency for each ALT allele in the same order as listed (AF).

TotalAlleleCount

Total number of alleles in called genotypes (AN).

BaseQuality

RMS base quality (BQ).

Cigar

Cigar string describing how to align an alternate allele to the reference allele (CIGAR).

IsInDbSnp

dbSNP membership (DB).

TotalDepth

Combined depth across samples (DP).

IsInHapMap2

HapMap2 membership (H2).

IsInHapMap3

HapMap3 membership (H3).

MappingQuality

RMS mapping quality (MQ).

ZeroMappingQualityCount

Number of MAPQ == 0 reads (MQ0).

SamplesWithDataCount

Number of samples with data (NS).

StrandBias

Strand bias (SB).

IsSomaticMutation

Somatic mutation (SOMATIC).

IsValidated

Validated by follow-up experiment (VALIDATED).

IsIn1000Genomes

1000 Genomes membership (1000G).

IsImprecise

Imprecise structural variation (IMPRECISE).

IsNovel

Indicates a novel structural variation (NOVEL).

EndPosition

End position of the variant described in this record (END).

SvType

Type of structural variant (SVTYPE).

SvLengths

Difference in length between REF and ALT alleles (SVLEN).

PositionConfidenceIntervals

Confidence interval around POS for imprecise variants (CIPOS).

EndConfidenceIntervals

Confidence interval around END for imprecise variants (CIEND).

MicrohomologyLengths

Length of base pair identical micro-homology at event breakpoints (HOMLEN).

MicrohomologySequences

Sequence of base pair identical micro-homology at event breakpoints (HOMSEQ).

BreakpointIds

ID of the assembled alternate allele in the assembly file (BKPTID).

MobileElementInfo

Mobile element info of the form NAME,START,END,POLARITY (MEINFO).

MobileElementTransductionInfo

Mobile element transduction info of the form CHR,START,END,POLARITY (METRANS).

DbvId

ID of this element in Database of Genomic Variation (DBVID).

DbVarId

ID of this element in DBVAR (DBVARID).

DbRipId

ID of this element in DBRIP (DBRIPID).

MateBreakendIds

ID of mate breakends (MATEID).

PartnerBreakendId

ID of partner breakend (PARID).

BreakendEventId

ID of event associated to breakend (EVENT).

BreakendConfidenceIntervals

Confidence interval around the inserted material between breakends (CILEN).

AdjacentReadDepths

Read Depth of adjacency (DPADJ).

BreakendCopyNumber

Copy number of segment containing breakend (CN).

AdjacentCopyNumber

Copy number of adjacency (CNADJ).

CopyNumberConfidenceIntervals

Confidence interval around copy number for the segment (CICN).

AdjacentCopyNumberConfidenceIntervals

Confidence interval around copy number for the adjacency (CICNADJ).

Other(String)

Any other non-reserved key.