Module noodles_vcf::record::alternate_bases::allele::symbol

VCF record alternate bases allele symbol and structural variant.

Re-exports

pub use self::structural_variant::StructuralVariant;

Modules

structural_variant

VCF record alternate bases allele structural variant symbol and type.

Enums

ParseError

An error returned when a raw VCF record alternate base allele symbol fails to parse.

Symbol

A VCF alternate bases allele symbol.