Struct noodles_vcf::record::genotypes::genotype::Genotype

pub struct Genotype(_);

A VCF record genotype.

Implementations

impl Genotype

pub fn parse(s: &str, formats: &Formats, keys: &Keys) -> Result<Self, ParseError>

Parses a raw genotype for the given genotype keys.

Examples

use noodles_vcf::{
    self as vcf,
    header::{format::Key, Format},
    record::genotypes::{genotype::{field::Value, Field}, Genotype},
};

let header = vcf::Header::builder()
    .add_format(Format::from(Key::Genotype))
    .add_format(Format::from(Key::ConditionalGenotypeQuality))
    .build();

let keys = "GT:GQ".parse()?;

assert_eq!(
    Genotype::parse("0|0:13", header.formats(), &keys),
    Ok(Genotype::try_from(vec![
        Field::new(Key::Genotype, Some(Value::String(String::from("0|0")))),
        Field::new(Key::ConditionalGenotypeQuality, Some(Value::Integer(13))),
    ])?)
);

pub fn genotype(&self) -> Option<Result<Genotype, GenotypeError>>

Returns the VCF record genotypes genotype value.

This is a convenience method to return a parsed version of the genotype (GT) field value. Use [Self::get] with [Key::Genotype] to get the raw value.

Examples

use noodles_vcf::{
    self as vcf,
    header::{format::Key, Format},
    record::genotypes::{genotype::{field::Value, Field}, Genotype},
};

let header = vcf::Header::builder()
    .add_format(Format::from(Key::Genotype))
    .add_format(Format::from(Key::ConditionalGenotypeQuality))
    .build();

let keys = "GT:GQ".parse()?;

let genotype = Genotype::parse("0|0:13", header.formats(), &keys)?;
assert_eq!(genotype.genotype(), Some(Ok("0|0".parse()?)));